Convert factor_cylindrical to REAl when used for postprocessing. Fix diagnostics bug: advance by half a time step was not done properly in cylindrical and used rotated vectors. This was wrong as we need to work in the same local coordinate system

+++ src/pic2d_ElectronMoments.f90
@@ -156,10 +156,10 @@ SUBROUTINE COLLECT_ELECTRON_MOMENTS
      vijp1 = ax_i   * ay_jp1
      vip1jp1 = 1.0 - vij - vip1j - vijp1
 
-     rbufer_n(pos_i_j)     = rbufer_n(pos_i_j)     + vij*factor_cyl_vol(i)     !ax_i   * ay_j
-     rbufer_n(pos_ip1_j)   = rbufer_n(pos_ip1_j)   + vip1j*factor_cyl_vol(i+1)   !ax_ip1 * ay_j
-     rbufer_n(pos_i_jp1)   = rbufer_n(pos_i_jp1)   + vijp1*factor_cyl_vol(i)   !ax_i   * ay_jp1
-     rbufer_n(pos_ip1_jp1) = rbufer_n(pos_ip1_jp1) + vip1jp1*factor_cyl_vol(i+1) !ax_ip1 * ay_jp1
+     rbufer_n(pos_i_j)     = rbufer_n(pos_i_j)     + vij*REAL(factor_cyl_vol(i))     !ax_i   * ay_j
+     rbufer_n(pos_ip1_j)   = rbufer_n(pos_ip1_j)   + vip1j*REAL(factor_cyl_vol(i+1))   !ax_ip1 * ay_j
+     rbufer_n(pos_i_jp1)   = rbufer_n(pos_i_jp1)   + vijp1*REAL(factor_cyl_vol(i))   !ax_i   * ay_jp1
+     rbufer_n(pos_ip1_jp1) = rbufer_n(pos_ip1_jp1) + vip1jp1*REAL(factor_cyl_vol(i+1)) !ax_ip1 * ay_jp1
 
 ! the snapshot is taken after the electric field was calculated but before the electrons were advanced
 ! therefore, while the electron coordinates correspond to time level n
@@ -360,8 +360,7 @@ SUBROUTINE COLLECT_ELECTRON_MOMENTS
   DO j = c_indx_y_min, c_indx_y_max
      DO i = c_indx_x_min, c_indx_x_max
         IF (cs_N(i,j).GT.1.0e-9) THEN    ! note this is small but not zero
-
-           inv_N = 1.0 / ( cs_N(i,j)/factor_cyl_vol(i) )
+           inv_N = 1.0 / ( cs_N(i,j)/REAL(factor_cyl_vol(i)) )
 
            cs_VX(i, j) = cs_VX(i, j) * inv_N
            cs_VY(i, j) = cs_VY(i, j) * inv_N
